Fix memory leak when playing haptic clip on iOS (#415)
* SDK-116: Use @autoreleasepool to ensure objects are freed properly

The callbacks here are called from a thread created in Rust. Such a thread
does not contain an autorelease pool, and therefore any call to autorelease()
will not have an effect and leak memory. Internally in Core Haptics, multiple
calls to autorelease() are made.
By adding @autorelease, the objects release with autorelease() will be properly
released at the end of the block.

This fixes a large memory leak when playing haptics on iOS.

* Unity: Clarify that our API needs to be called from the main thread to avoid leaking