Add build.sh to create NiceVibrations asset folder
This script builds the SDK for all supported platforms. In other
words, it calls `build-sdk.sh` to build the iOS Framework, Android
library and Unity Editor plugin host.
The Unity Editor plugin host is only build for the platform where
`build.sh` is called from. This means there is no cross-compilation
for macOS and Windows.

The script creates a `nice-vibrations-asset` containing a
NiceVibrations asset folder that can be imported into a Unity project.